// raw/sem_resources_generic/06/solution-06/materials-feat-seminar-06-week-06-seminar-solution/week-06/seminar/solution/apps/web/src/generated/types/Course.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const courseSemesterEnum = {
    "fall": "fall",
    "spring": "spring"
} as const;

export type CourseSemesterEnum = (typeof courseSemesterEnum)[keyof typeof courseSemesterEnum];

export type Course = {
    /**
     * @type string
    */
    id: string;
    /**
     * @type string
    */
    code: string;
    /**
     * @type string
    */
    name: string;
    /**
     * @type string
    */
    description: string;
    /**
     * @minLength -9007199254740991
     * @maxLength 9007199254740991
     * @type integer
    */
    credits: number;
    /**
     * @type string
    */
    instructorId: string;
    /**
     * @type string
    */
    semester: CourseSemesterEnum;
    /**
     * @minLength -9007199254740991
     * @maxLength 9007199254740991
     * @type integer
    */
    year: number;
    /**
     * @minLength -9007199254740991
     * @maxLength 9007199254740991
     * @type integer
    */
    capacity: number;
    /**
     * @minLength -9007199254740991
     * @maxLength 9007199254740991
     * @type integer
    */
    enrolled: number;
};